Manager and store

+++ shared/database/src/commonMain/kotlin/com/sedsoftware/tackle/database/model/CachedConfig.kt
@@ -38,8 +38,8 @@ internal class CachedConfig(
 
     @Serializable
     class CachedStatuses(
-        val maxCharacters: Long = 0L,
-        val maxMediaAttachments: Long = 0L,
-        val charactersReservedPerUrl: Long = 0L,
+        val maxCharacters: Int = 0,
+        val maxMediaAttachments: Int = 0,
+        val charactersReservedPerUrl: Int = 0,
     )
 }
