+++ shared/domain/src/commonMain/kotlin/com/sedsoftware/tackle/domain/model/Instance.kt
@@ -47,9 +47,9 @@ data class Instance(
         )
 
         class Statuses(
-            val maxCharacters: Long = 0L,
-            val maxMediaAttachments: Long = 0L,
-            val charactersReservedPerUrl: Long = 0L,
+            val maxCharacters: Int = 0,
+            val maxMediaAttachments: Int = 0,
+            val charactersReservedPerUrl: Int = 0,
         )
     }
